Fix mappers behavior with foreign keys and enums
Fixes #3

Fix foreign keys behavior: now, when a foreign key is detected, the
foreign object will be fetched and linked as a new mapper. Black magic
is done with properties to fetch this object in a JIT way: accessing the
object for the first time will do the request to fetch it, then it will
be stored in cache local to the parent.

Fix serialization of netbox choices: not ideal for now, could be
improved later, but at least works.

Handle specific cases when a mapper cannot be generated.